+++ Runtime/Animations/AbstractAnimation.cs
@@ -4,6 +4,7 @@ namespace PrimeTween
 {
     public abstract class AbstractAnimation : MonoBehaviour
     {
+        public abstract Awaitable PlayAsync();
         public abstract void Play();
         public abstract void Stop();
     }