+++ Runtime/Animations/AnimationSequence.cs
@@ -9,6 +9,14 @@ namespace PrimeTween
 
         private void Awake() => animations = GetComponentsInChildren<AbstractAnimation>();
 
+        public override async Awaitable PlayAsync()
+        {
+            foreach (var animation in animations)
+            {
+                await animation.PlayAsync();
+            }
+        }
+
         public override void Play()
         {
             foreach (var animation in animations)