+++ changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Async AbstractAnimation.PlayAsync function
+
 ### Changed
 - Rename package into Action Code-Prime Tween
 - Update original packaged into 1.3.8